pyproject.toml
[tool.poetry]
name = "intrepyx"
version = "0.1.0"
description = ""
authors = ["Jean-Luc Tromparent <jlt@kezako.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-socketio = {extras = ["client"], version = "^5.8.0"}
pydantic = "^1.10.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "intrepyx.main:start"